Web Packages (AP3, WO,FW,DL): if(_slack.isSelected()) check before send report(s) to Slack. Was missing in previous commit

+++ src/main/java/AP3/AP3.java
@@ -1268,18 +1268,20 @@ public class AP3 extends javax.swing.JInternalFrame {
                 
                 LOG_UPDATE(); // ========================================================
                 
-                Report(false);
-                
-                String MSG = "AP3_" + env + " Automation report - " + Report_Date +  
-                "\r\n Machine: " + WsID + " OS: " + WsOS + ", User: *" + UserID + "*\r\n" +
-                "Duration: " + DD.toHours() + "h, " + (DD.toMinutes() % 60) + "m, " + (DD.getSeconds() % 60) + "s" + "\r\n" +        
-                "Scope: " + SCOPE + "\r\n" +
-                "Steps: " + _t + ", Passed: " + _p + ", *Failed: " + _f + "*, Warnings: " + _w;
-                
-                txtLog.append(Func.Send_File_to_Slack(Report_File, "ap3automation", MSG));
-                File f = new File(Report_File);
-                if(f.exists() && !f.isDirectory()) { 
-                    f.delete();
+
+                if(_slack.isSelected()){
+                    Report(false); 
+                    String MSG = "AP3_" + env + " Automation report - " + Report_Date +  
+                    "\r\n Machine: " + WsID + " OS: " + WsOS + ", User: *" + UserID + "*\r\n" +
+                    "Duration: " + DD.toHours() + "h, " + (DD.toMinutes() % 60) + "m, " + (DD.getSeconds() % 60) + "s" + "\r\n" +        
+                    "Scope: " + SCOPE + "\r\n" +
+                    "Steps: " + _t + ", Passed: " + _p + ", *Failed: " + _f + "*, Warnings: " + _w;
+
+                    txtLog.append(Func.Send_File_to_Slack(Report_File, "ap3automation", MSG));
+                    File f = new File(Report_File);
+                    if(f.exists() && !f.isDirectory()) { 
+                        f.delete();
+                    }
                 }
             } 
         };
